Moved functions to transforms from topics.
