+++ src/topics/property.js
@@ -1,11 +1,7 @@
 import helpers from '../util/helpers';
+import transforms from '../general/transforms';
 
-const titleCase = function(str) {
-  str = str.toLowerCase().split(' ').map(function(word) {
-    return (word.charAt(0).toUpperCase() + word.slice(1));
-  });
-  return str.join(' ');
-};
+const titleCase = transforms.titleCase.transform;
 
 export default {
   key: 'property',
