+++ src/topics/trash.js
@@ -1,12 +1,5 @@
 import transforms from '../general/transforms';
-
-const getNearest = function(state, field, distName) {
-
-  let min = Math.min.apply(null, state.sources[field].data.map(function(item) {return item[distName];}));
-  let result  = state.sources[field].data.filter(function(item){return item[distName] == min;} );
-  let nearest = result? result[0] : null;
-  return nearest
-};
+const getNearest = transforms.getNearest.transform;
 
 
 export default {
