+++ src/topics/voting.js
@@ -1,22 +1,13 @@
 import transforms from '../general/transforms';
 import moment from 'moment';
 const phone = transforms.phoneNumber.transform;
-
-const titleCase = function(str) {
-  str = str.toLowerCase().split(' ').map(function(word) {
-    return (word.charAt(0).toUpperCase() + word.slice(1));
-  });
-  return str.join(' ');
-};
-
-const nth = function(n) {return n + ([,'st','nd','rd'][n%100>>3^1&&n%10]||'th')};
-
+const titleCase = transforms.titleCase.transform;
+const nth = transforms.nth.transform;
 
 export default {
   key: 'voting',
   icon: 'gavel',
   label: 'Voting',
-  dataSources: ['divisions', 'pollingPlaces', 'electedOfficials'],
   dataSources: ['divisions', 'pollingPlaces', 'electedOfficials', 'nextElectionAPI'],
   components: [
     {
